Started separating TDAmeritrade API related functions in the tda namespace into their own files and implemented those changes in the build process.
