+++ src/TDAmeritrade/Data/Quote.cpp
@@ -1,0 +1,26 @@
+#include "Quote.hpp"
+
+/* =============== Quote Class =============== */
+
+void tda::Quote::initVariables()
+{
+    for (auto & array_element: quoteData ) 
+    {
+        for (auto & property: array_element.second) 
+        {
+            quoteVariables[ property.first ] = property.second.get_value < std::string > ();
+        }
+    }
+}
+
+tda::Quote::Quote( boost::property_tree::ptree quote_data )
+{
+    quoteData = quote_data;
+
+    initVariables();
+}
+
+std::string tda::Quote::getQuoteVariable( std::string variable )
+{
+    return quoteVariables[variable];
+}
