+++ src/TDAmeritrade/Data/Quote.hpp
@@ -1,0 +1,24 @@
+#ifndef Quote_hpp
+#define Quote_hpp
+
+#include "../../core.hpp"
+
+namespace tda 
+{
+    class Quote 
+    {
+    private:
+        boost::property_tree::ptree quoteData;
+        std::map<std::string, std::string> quoteVariables;
+
+        void initVariables();
+
+    public:
+        Quote( boost::property_tree::ptree quote_data );
+
+        std::string getQuoteVariable( std::string variable );
+
+    };
+}
+
+#endif
